requirements/base.txt

# Django
django==4.2.7
django-environ==0.11.2
django-mptt==0.15.0

psycopg2-binary==2.9.9

# third party libraries
django-redis==5.4.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
celery==5.3.6
Pillow==10.1.0
elasticsearch-dsl==8.11.0
django-elasticsearch-dsl==8.0
# Django REST Framework
djangorestframework==3.14.0
django-cors-headers==4.3.0
djangorestframework_simplejwt==5.3.0
# DRF-spectacular for api documentation
drf-spectacular==0.26.5  # https://github.com/tfranzel/drf-spectacular
